Add XML Documentation to `IPasswordlessClient` (#16)
* Add Docs

- Convert List responses to IReadOnlyList
- Bump version to 0.2.0

* Make ListResponse Internal

* Format

---------
